chore: move RPC functions
